refactor: add Container component
