+++ utils/index.js
@@ -1,5 +1,3 @@
-import ReactSelect from 'react-select';
-import styled from 'styled-components';
 import Yup from 'yup';
 import { sectionListOptions } from './options';
 
@@ -66,12 +64,3 @@ export const schema = Yup.object().shape({
     value: Yup.number().required()
   })
 });
-
-export const Select = styled(ReactSelect)`
-  &.Select.error > .Select-control {
-    border-color: #dc3545;
-    box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.075),
-      0 0 0 3px rgba(220, 53, 69, 0.1);
-    background: #fff;
-  }
-`;
